Made sifs.find and sil.find more flexible regarding search patterns.
The main motivation was the desire to specify a subdirectory
for a sil file I was using:
  sil php/migrations
didn't work with the old sil.find (using the -iname switch).
If you have sil files that have the same name, just
qualify with one or more parent directories like the
above example and it seems to work quite well.
eg
  sil rails/migrations
  sil php/migrations